+++ sample/gen_protocol_v3.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-import os, sys
+import os
 import string
 import glob
 
@@ -12,7 +12,7 @@ os.chdir(script_dir)
 os.chdir(os.path.join('..'));
 
 project_dir = os.getcwd();
-proto_dir = os.path.join(script_dir, 'proto_v2');
+proto_dir = os.path.join(script_dir, 'proto_v3');
 
 proto_file = []
 for item in glob.glob(os.path.join(proto_dir, '*.proto')):
@@ -22,10 +22,10 @@ os.chdir(work_dir);
 
 cpp_out_dir = os.path.join(script_dir, 'cxx');
 
-proto_src_dir = '{0}/v2'.format(cpp_out_dir)
+proto_src_dir = '{0}/v3'.format(cpp_out_dir)
 if not os.path.exists(proto_src_dir):
     os.mkdir(proto_src_dir)
-params = ['protoc', '-I', proto_dir, '-o', os.path.join(proto_dir, 'kind.pb'), '--cpp_out={0}'.format(proto_src_dir)]
+params = ['protoc', '-I', proto_dir, '-o', os.path.join(proto_dir, 'kind.pb'), '--cpp_out=lite:{0}'.format(proto_src_dir)]
 params.extend(proto_file)
 cmd = ' '.join(params)
 print(cmd)