+++ tools/gen_header_v3.py
@@ -13,18 +13,14 @@ os.chdir(os.path.join('..'));
 
 project_dir = os.getcwd();
 proto_dir = os.path.join(project_dir, 'header');
-proto_file = os.path.join(proto_dir, 'pb_header.proto');
+proto_file = os.path.join(proto_dir, 'pb_header_v3.proto');
 
 os.chdir(work_dir);
 
 java_out_dir = proto_dir;
-pb_out_file = os.path.join(proto_dir, 'pb_header.pb');
+pb_out_file = os.path.join(proto_dir, 'pb_header_v3.pb');
 
 # java 文件为非LITE版本
-proto_file_data = open(proto_file).read()
-proto_file_no_lite_data = re.sub('\s*option\s*optimize_for\s*=\s*LITE_RUNTIME\s*;','// option optimize_for = LITE_RUNTIME;', proto_file_data)
-open(proto_file, 'wb').write(proto_file_no_lite_data.encode('utf-8'))
-
 params = ['protoc', '-I', '"' + proto_dir + '"' , '--java_out',  '"' + java_out_dir + '"', '"' + proto_file + '"']
 cmd = ' '.join(params)
 print('[PROCESS] generate java source ... ')
@@ -33,8 +29,6 @@ os.system(cmd)
 print('[PROCESS] generate java source done.')
 
 # pb 文件为LITE版本
-open(proto_file, 'wb').write(proto_file_data.encode('utf-8'))
-
 params = ['protoc', '-I', '"' + proto_dir + '"' , '-o',  '"' + pb_out_file + '"', '"' + proto_file + '"']
 cmd = ' '.join(params)
 print('[PROCESS] generate proto pb file ... ')